Penambahan lihat surat perjalanan dinas di LHP

+++ application/views/administrator/datajabatan_admin.php
@@ -1,4 +1,24 @@
-<font face="calibri" size = "6" color="black">Selamat Datang <?php echo $_SESSION['nama'] ;?></font>
+<center><div class="alert alert-success alert-dismissible" style = "color: blue;" role="alert"><h4>
+               Selamat datang <?php echo $_SESSION['nama']?>, anda login sebagai</strong>
+                <?php
+                switch ($_SESSION['akses']) {
+                  case 'tu':
+                    echo 'Tata Usaha';
+                    break;
+                  case 'admin':
+                    echo 'Admin';
+                    break;
+                     case 'pegawai ':
+                    echo 'Pegawai';
+                    break;
+
+                    default:
+                    # code...
+                    break;
+                }
+              ?>
+          <strong><h4>
+      </div> </center>
 <div class="row">
         <div class="col-xs-12">
 			<div class="box">
@@ -7,26 +27,26 @@
             </div>
             <!-- /.box-header -->
             <div class="box-body">
-               <a href="<?php echo base_url()."index.php/Administrator/Administrator/input_jabatan"?>" button type="button" class="btn btn-primary float-left"><i class="fa fa-plus-square-o"></i> Tambah </a>
+               <a href="<?php echo base_url()."administrator/administrator/input_jabatan"?>" button type="button" class="btn btn-primary float-left"><i class="fa fa-plus-square-o"></i> Tambah </a>
                <hr>
-
+ <?php if ($this->session->pesan) echo $this->session->pesan; ?> 
               <table id="example1" class = "table table-bordered table-striped">
                 <thead>
                 <tr class="nowrap">
-                  <th>ID</th>
+                  <th>No</th>
                   <th>Jabatan</th>  
                   <th>Aksi</th>
                 </tr>
                 </thead>
                 <tbody>
-                	<?php foreach ($data->result() as $row) : ?>
+                	<?php     $no = 1; foreach ($data->result() as $row) : ?>
                 <tr>
-                  <td><?php echo $row->id?></td>
+                  <td><?php echo $no++?></td>
                   <td><?php echo $row->jabatan?></td>
                   <td>
                     <a href="edit_jabatan/<?php echo $row->id ?>" type="button" class="btn btn-success btn-sm" ><i class="fa fa-edit"></i> Ubah </a>
                
-                    <a href="hapus_jabatan/<?php echo $row->id ?>" type="button" class="btn btn-danger btn-sm" ><i class="fa fa-trash"></i> Hapus </a>
+                    <a href="hapus_jabatan/<?php echo $row->id ?>" type="button" class="btn btn-danger btn-sm" onClick="return confirm('Apakah anda yakin ingin menghapus data jabatan dengan id : <?php echo $row->id ?>');" ><i class="fa fa-trash"></i> Hapus </a>
 
                   </td>
